+++ lib/go-atscfg/regexrevalidatedotconfig_test.go
@@ -64,6 +64,15 @@ func TestMakeRegexRevalidateDotConfig(t *testing.T) {
 			Parameters:      "TTL:14h",
 			Keyword:         JobKeywordPurge,
 		},
+		tc.Job{
+			AssetURL:        "refetchasset__REFETCH__",
+			StartTime:       time.Now().Add(24 * time.Hour).Format(tc.JobTimeFormat),
+			DeliveryService: "myds",
+			CreatedBy:       "want_refetch",
+			ID:              42,
+			Parameters:      "TTL:24h",
+			Keyword:         JobKeywordPurge,
+		},
 	}
 
 	cfg, err := MakeRegexRevalidateDotConfig(server, dses, params, jobs, hdr)
@@ -80,5 +89,8 @@ func TestMakeRegexRevalidateDotConfig(t *testing.T) {
 	}
 	if strings.Contains(txt, "expired") {
 		t.Errorf("expected no expired job, actual '%v'", txt)
+	}
+	if strings.Contains(txt, "__REFETCH__") || !strings.Contains(txt, "MISS") {
+		t.Errorf("__REFETCH__ directive not properly handled '%v'", txt)
 	}
 }